Redo csc calculations
There is no possibility to forward the csc matrix to linux-sunxi kernel
display driver directly with VDPAU. We have to set the values for
brightness, saturation, contrast and hue via ioctl. But there are no
attributes for these values in VDOAU to get them directly. So we need
to do some reverse calculation to get these values out of the (changed)
csc matrix.
The old calculation was wrong and is now corrected to get the right
values for brightness, saturation, contrast and hue. Due to the reverse
calculation algorithm it's not possible to get all values for every
scenario. This usecase is documented in the source code.

Note:
The linux-sunxi 3.4 kernel is buggy when doing this calculations. So
in order to get this work correctly, you have to patch it or use an
already patched kernel.