Detect different *render_surface operations
When we have done a blitting operation, we save the parameters
of this blit into the destination rgba surface.
The next time, some surface wants to be blitted, we check,
if something changed at all (e.g. new id during put_bits_*,
different rects). If nothing changed, compared to the saved last
blit, we simply do nothing, but mark the surface as dirty and
display it.

Dropping the blit gives us some performance boost.