Add caching for rgba surfaces
This commit adds caching for rgba surfaces. We don't
create an rgba surface during output/surface creation
anymore. Instead we now try to get a free rgba surface
from a cache directly before a render or put_bits_*-
action is triggered.

This feature also includes a check, if we plan to do a
new/ different render action at all. Only in the case
that we have some changes at all, we do a real blit.
If there is no change at all, we just do a zero-copy.
This commit should also take care, that no other
surfaces (which hold a pointer to the same rgba) are
affected by counting the rgba's references.

As the blitting operation within the render_*_surface
command could be kind of a bottleneck, this commit
tries to calm them down to the minimum.