Detect different output_surface_render_*_surface operations
When a blitting operation is done, save the parameters of
this blit into the destination rgba_surface.
The next time, when some surface wants to be blitted, just
check if something changed at all (e.g. the source has a new
id because of a new put_bit_* action in the meantime, or
different rects). If nothing changed, compared to the
saved last blit, we simply do nothing, skip the blit,
mark the surface as dirty and display it.